requirements.txt
click==8.1.7
ConfigArgParse==1.7
cv_bridge==1.16.2
dm_control==1.0.16
dm_env==1.6
einops==0.8.0
gym==0.13.0
h5py==3.7.0
ipython==8.12.3
matplotlib==3.8.0
mj_envs.egg==info
mjrl.egg==info
moviepy==1.0.3
mujoco_py==2.1.2.14
numpy==2.0.0
opencv_python==4.8.1.78
openvr==1.26.701
optuna==3.6.1
pandas==2.2.2
Pillow==10.0.1
Pillow==10.4.0
pykalman==0.9.7
pyquaternion==0.9.9
pytorch_lightning==2.1.3
PyYAML==6.0.1
PyYAML==6.0.1
quaternion==3.5.2.post4
robohive.egg==info
roslib==1.15.8
rospy==1.16.0
scipy==1.14.0
seaborn==0.13.2
sensor_msgs==1.13.1
soundfile==0.12.1
tabulate==0.9.0
tensorflow==2.14.0
tf==1.13.2
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transforms3d==0.4.1
